Extracted SearchField from SearchBar component to its own component
